rename math packages
