+++ apps/balancer-tools/package.json
@@ -18,8 +18,8 @@
   "dependencies": {
     "@balancer-labs/sor": "4.1.1-beta.13",
     "@bleu-balancer-tools/gql": "workspace:*",
-    "@bleu-balancer-tools/math": "workspace:*",
-    "@bleu-balancer-tools/math-new": "workspace:*",
+    "@bleu-balancer-tools/math-stableswapsimulator": "workspace:*",
+    "@bleu-balancer-tools/math-poolsimulator": "workspace:*",
     "@bleu-balancer-tools/tsconfig": "workspace:*",
     "@hookform/resolvers": "^3.1.1",
     "@radix-ui/colors": "^1.0.0",
@@ -61,7 +61,7 @@
   },
   "devDependencies": {
     "@bleu-balancer-tools/eslint-config": "workspace:^",
-    "@bleu-balancer-tools/math": "workspace:^",
+    "@bleu-balancer-tools/math-stableswapsimulator": "workspace:^",
     "@bleu-balancer-tools/utils": "workspace:^",
     "@ethersproject/bignumber": "^5.7.0",
     "@testing-library/jest-dom": "5.17.0",
