+++ packages/math-poolsimulator/README.md
