+++ packages/math-poolsimulator/jest.config.js
