+++ packages/math-poolsimulator/src/index.ts
