+++ packages/math-poolsimulator/tsconfig.json
