+++ packages/math-stableswapsimulator/jest.config.js
