+++ packages/math-stableswapsimulator/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@bleu-balancer-tools/math-new",
+  "name": "@bleu-balancer-tools/math-stableswapsimulator",
   "version": "0.0.0",
   "main": "index.ts",
   "license": "MIT",
@@ -13,13 +13,11 @@
   "dependencies": {
     "@balancer-labs/sor": "4.1.1-beta.9",
     "@ethersproject/bignumber": "^5.7.0",
-    "@ethersproject/constants": "^5.7.0",
-    "bignumber.js": "^9.1.1"
+    "@ethersproject/constants": "^5.7.0"
   },
   "devDependencies": {
     "@bleu-balancer-tools/tsconfig": "workspace:^",
-    "@jest/globals": "^29.5.0",
-    "ts-jest": "^29.1.0",
-    "tsuml2": "^0.12.0"
+    "@jest/globals": "^29.6.1",
+    "ts-jest": "^29.1.1"
   }
 }
