+++ packages/math-stableswapsimulator/src/index.ts
