+++ packages/math-stableswapsimulator/src/metastable/index.ts
