+++ packages/math-stableswapsimulator/src/stable/fixtures/data.json
