+++ packages/math-stableswapsimulator/src/stable/index.test.ts
