+++ packages/math-stableswapsimulator/src/stable/index.ts
