+++ packages/math-stableswapsimulator/tsconfig.json
